[iOS] feat: CoreTrackAPI 생성 및 TrackCellViewModel에 적용
